+++ apps/ledger-live-mobile/src/components/RootNavigator/NotificationCenterNavigator.tsx
@@ -1,25 +1,20 @@
-import React, { useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
-import { useSelector } from "react-redux";
 import NotificationCenterStatus from "../../screens/NotificationCenter/Status";
-import NotificationCenterNews from "../../screens/NotificationCenter/News";
+import NotificationCenterNews from "../../screens/NotificationCenter/Notifications";
 import { ScreenName } from "../../const";
 import type { NotificationCenterNavigatorParamList } from "./types/NotificationCenterNavigator";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
-import { notificationsCardsSelector } from "../../reducers/dynamicContent";
 
 const Stack = createStackNavigator<NotificationCenterNavigatorParamList>();
 
 export default function NotificationCenterNavigator() {
   const { t } = useTranslation();
   const { colors } = useTheme();
-  const notifications = useSelector(notificationsCardsSelector);
-  const [notificationsCount] = useState(
-    notifications.length - notifications.filter(n => n.viewed).length,
-  );
+
   const stackNavConfig = useMemo(
     () => getStackNavigatorConfig(colors),
     [colors],
@@ -31,14 +26,7 @@ export default function NotificationCenterNavigator() {
         name={ScreenName.NotificationCenterNews}
         component={NotificationCenterNews}
         options={{
-          title: t(
-            notificationsCount > 0
-              ? "notificationCenter.news.titleCount"
-              : "notificationCenter.news.title",
-            {
-              count: notificationsCount,
-            },
-          ),
+          title: t("notificationCenter.news.title"),
         }}
       />
       <Stack.Screen
